requirements.txt
Flask
google-cloud-vision
gunicorn
google-generativeai
Pillow